requirements.txt
openai
torch
pdfplumber
streamlit
xformers
transformers
tiktoken
pandas
langchain
sentence-transformers
faiss-cpu
langchain-community
pymilvus==2.4.12
langchain-huggingface
python-docx
google-cloud-secret-manager
